Removed arguments from Exec call
One goal I have with this package is to be able to write command line
commands that can be executed independently or from within the shell.
Since independent commands will parse their arguments from os.Args the
gosh package now sets os.Args to the args (with Args[0] being the
command name) and leaves it up to the implementor to retrieve the
arguments from that string slice.  This allows the Command implementor
to use things like the flag package from within gosh